+++ packages/auth/src/types/Auth.ts
@@ -76,13 +76,18 @@ export interface FederatedUser {
     attributes?: Object
 }
 
-export interface OAuth {
-    domain : string,
-	scope : Array<string>,
-	redirectSignIn : string,
-	redirectSignOut : string,
+export interface awsCognitoOAuthOpts {
+    domain: string,
+	scope: Array<string>,
+	redirectSignIn: string,
+	redirectSignOut: string,
     responseType: string,
     options?: object
+}
+
+export interface OAuth {
+    awsCognito?: awsCognitoOAuthOpts,
+    auth0?: any
 }
 
 export interface ConfirmSignUpOptions {