+++ packages/aws-amplify-react/src/Storage/S3Image.js
@@ -11,7 +11,8 @@
  * and limitations under the License.
  */
 
-import React, { Component } from 'react';
+import * as React from 'react';
+import { Component } from 'react';
 
 import { ConsoleLogger as Logger } from '@aws-amplify/core';
 import Storage from '@aws-amplify/storage';
@@ -145,7 +146,7 @@ export default class S3Image extends Component {
                 />
                 <div style={selected? theme.overlaySelected : theme.overlay}></div>
             </div>
-        )
+        );
     }
 
     render() {
@@ -176,6 +177,6 @@ export default class S3Image extends Component {
                           </div> : null
                 }
             </div>
-        )
+        );
     }
 }