+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.2.4"></a>
+## [1.2.4](https://github.com/aws/aws-amplify/compare/@aws-amplify/analytics@1.2.4-unstable.4...@aws-amplify/analytics@1.2.4) (2018-10-17)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+<a name="1.2.4-unstable.4"></a>
+## [1.2.4-unstable.4](https://github.com/aws/aws-amplify/compare/@aws-amplify/analytics@1.2.4-unstable.3...@aws-amplify/analytics@1.2.4-unstable.4) (2018-10-17)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+<a name="1.2.4-unstable.3"></a>
+## [1.2.4-unstable.3](https://github.com/aws/aws-amplify/compare/@aws-amplify/analytics@1.2.4-unstable.2...@aws-amplify/analytics@1.2.4-unstable.3) (2018-10-16)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
 <a name="1.2.4-unstable.2"></a>
 ## [1.2.4-unstable.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/analytics@1.2.4-unstable.1...@aws-amplify/analytics@1.2.4-unstable.2) (2018-10-08)