+++ packages/xr/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.6"></a>
+## [0.1.6](https://github.com/aws/aws-amplify/compare/@aws-amplify/xr@0.1.6-unstable.4...@aws-amplify/xr@0.1.6) (2018-10-17)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/xr
+
+<a name="0.1.6-unstable.4"></a>
+## [0.1.6-unstable.4](https://github.com/aws/aws-amplify/compare/@aws-amplify/xr@0.1.6-unstable.3...@aws-amplify/xr@0.1.6-unstable.4) (2018-10-17)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/xr
+
+<a name="0.1.6-unstable.3"></a>
+## [0.1.6-unstable.3](https://github.com/aws/aws-amplify/compare/@aws-amplify/xr@0.1.6-unstable.2...@aws-amplify/xr@0.1.6-unstable.3) (2018-10-16)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/xr
+
 <a name="0.1.6-unstable.2"></a>
 ## [0.1.6-unstable.2](https://github.com/aws/aws-amplify/compare/@aws-amplify/xr@0.1.6-unstable.1...@aws-amplify/xr@0.1.6-unstable.2) (2018-10-08)